fix: disable image optimization for edge
