+++ app/layout.js
@@ -1,26 +1,28 @@
-"use client";
 import "./globals.css";
 import Theming from "@/components/providers/Theme";
-import Dashboard from "@/app/dashboard/layout";
-export const runtime = 'edge';
+
+export const runtime = "edge";
 
 export default function RootLayout({ children }) {
   return (
-    <html lang="hu">
+    <html lang="en">
       <head>
-        <title>PromNET - Polyák Csaba E.V.</title>
-        <meta name="description" content="Weboldal fejlesztés gyorsan, olcsón!" />
-        <meta property="og:title" content="PromNET - Polyák Csaba E.V." />
-        <meta property="og:description" content="Weboldal fejlesztés gyorsan, olcsón!" />
+        <title>PromNET - Csaba Polyak</title>
+        <meta
+          name="description"
+          content="Website development quickly and affordably!"
+        />
+        <meta property="og:title" content="PromNET - Csaba Polyak" />
+        <meta
+          property="og:description"
+          content="Website development quickly and affordably!"
+        />
         <meta property="og:image" content="/logo-white.png" />
         <meta property="og:url" content="https://www.promnet.hu" />
         <meta name="twitter:card" content="summary_large_image" />
       </head>
       <body className="bg-[#171717]">
-        <Theming>
-          { <Dashboard /> }
-          {children}
-        </Theming>
+        <Theming>{children}</Theming>
       </body>
     </html>
   );
