+++ next.config.js
@@ -1,11 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  /* config options here */
+  images: {
+    // Disable Next.js image optimization for Cloudflare compatibility
+    unoptimized: true,
+  },
 };
 
 module.exports = nextConfig;
-module.exports = {
-  images: {
-      unoptimized: true
-  }
-}
